+++ src/main/java/com/ops/ops/global/util/oauth/dto/GoogleOAuthToken.java
@@ -1,4 +1,4 @@
-package com.ops.ops.global.util.oauth.token;
+package com.ops.ops.global.util.oauth.dto;
 
 public record GoogleOAuthToken(
 	String access_token,
