+++ src/main/java/com/ops/ops/global/util/oauth/dto/GoogleUser.java
@@ -1,4 +1,4 @@
-package com.ops.ops.global.util.oauth.model;
+package com.ops.ops.global.util.oauth.dto;
 
 public record GoogleUser(
 	String email,
